chore: rename file - userModel -> authModel
파일 이름 변경
userModel -> authModel
